requirements.txt
requests==2.22.0
urllib3==1.24.2
pyTelegramBotAPI
telebot~=0.0.4
PyYAML~=6.0
expandvars